[HAO-Tracker-2014-Primary] Fixed a regression bug in main.cpp
- Since AnalogSensors API has been changed in order to use a pointer to
an AnalogSensor array instead of making a copy of an array coming as
argument, the constructor parameters have to be modified. Then, in
main.cpp, the constructor calls related the two AnalogSensors to be
created take
(AnalogSensor **) as parameter and consequently need pointers to
AnalogSensors arrays to be passed instead of simply AnalogSensors
arrays.