[HAO-Tracker-2014-Primary] Updated counters related code in main.cpp
- Since Counters API has been changed in order to use a pointer to
an Counter array instead of making a copy of an array coming as
argument, the constructor parameters have to be modified. Then, in
main.cpp, the constructor calls related to Counters to be
created take
(Counter **) as parameter and consequently need pointers to
counters arrays to be passed instead of simply counters
arrays.